[move] slideEvent.js: main.js에서 silde와 관련된 함수를 모듈로 분리함
