+++ JS/main.js
@@ -3,6 +3,7 @@ import {todayWebtoonsData} from './data/todayWebtoons.js'
 import {renderHome, renderDaily} from './render.js'
 import {dailyTopData} from './data/scrape/dailyTopData.js'
 import {makeWebtoonList} from './components/webtoonList.js'
+import {eventMainBanner} from './slideEvent.js'
 import {$, $all} from './utility.js'
 
 const setFocus = (target, className) => {
@@ -72,49 +73,6 @@ const selectDailyWebtoonData = (parentNode, data) => {
     : data.filter((el) => el.day.includes(select) && !(el.review.includes('위')));
 
   return filterData;
-}
-
-const makeCloneNode = () => {
-  const list = $('.main__banner .banner__list');
-  const items = $all('.main__banner .banner__item');
-  const firstNode = items[0].cloneNode(true);
-  const lastNode = items[items.length - 1].cloneNode(true);
-  firstNode.classList.add('cloneFirstNode');
-  lastNode.classList.add('cloneLastNode');
-
-  list.appendChild(firstNode);
-  list.insertBefore(lastNode, items[0]);
-
-  return list
-}
-
-const eventMainBanner = () => {
-  const list = makeCloneNode();
-  const items = $all('.main__banner .banner__item');
-  const width = $('.main__banner .banner__item').offsetWidth;
-
-  let counter = 1;
-  list.style.transform = `translate(${-width * counter}px)`;
-  slideMainBanner(list, items, width, counter);
-}
-
-const slideMainBanner = (list, items, width, counter) => {
-  setInterval(() => {
-    if (counter >= items.length - 1) return;
-    counter++;
-    list.style.transform = `translate(${-width * counter}px)`;
-    list.style.transition = 'transform 0.5s';
-
-    $('.banner__count .now').innerText = items[counter].dataset.index;
-  }, 3000);
-
-  list.addEventListener('transitionend', () => {
-    if (items[counter].classList.contains("cloneFirstNode"))  {
-      list.style.transition = "none";
-      counter = items.length - counter;
-      list.style.transform = `translate(${-width * counter}px)`;
-    }
-  })
 }
 
 const clickBannerPrevBtn = () => {
